requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyjwt[crypto]==2.8.0
cryptography==42.0.0
bcrypt==4.0.1
python-multipart==0.0.6
jinja2==3.1.3
lxml==5.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
